Replace old typename-array header with submodule.
Remove old directory and header, so that it can be replaced with a
submodule. That thing will be supported as a separate project, because
it became too big. Add a submodule for typename-array, this version of
it was also fixed and contains in-code documentation. Change includes
and namespace names accordingly, new version of typename-array should
work identically to the old one otherwise.