Start using std::osyncstream for logger.
Previously it just used a global lock to ensure that
exactly one thread is writing to std::cerr. It (cerr stream)
is already required to be thread-safe. However, I am pretty
sure that it does not guarantee that output won't be garbled,
so we must synchronize output to it.
std::osyncstream is prefect for this.