Introduce Visual Leak Detector for Debug builds.
Create a properties file which contains the base path for VLD.
Reconfigure all modules so that they have additional lib and header
paths for VLD. VLD is unconditionally disabled for Release (and
derivative builds). VLD is disabled by default for all builds, this is
because for some unknown reason it refuses to properly identify memory
leaks in DLLs if executable has it imported (or any other DLL, for that
matter). So, the only solution is to enable it on per-binary basis when
I want to test for memory leaks.

NOTE: It also seems that VLD reports memory leaks ONLY for the first
binary that imports it. Judging by its source, this should not be the
case because it modifies ntdll in such a manner that, when a new module
is loading, it should refresh its module list, attaching to new modules.
I suspect that there might be either an error on my side (However, I
know for sure that it works if exactly one binary is using it), an error
in reporting (Also unlikely because the function that reporst memory
leaks just accesses one global object), some error with trying to ensure
that there is only one global vld class object.