WIP: Standard IO support for PRTS.
Implement abstraction functions for accessing STDIN and STDOUT (with support for console, files, pipes). Implement start up and shutdown sequences for IO workers. Perfrom automated refactoring (this change just so happens to be included here). Create a deferred callback system for execution module to enable full support of thread blocking.

This is so complicated that I am not even sure whether it'll work. However, I can't even debug it yet because the system is incomplete.