Start working on "About" window. Fix parsing bug. Add additional messages.
Provide additional information when user tries to start a second instance of the FSI tool. For translator this is the last program used for translation, along with destination. For mediator this is the last started program, along with the configuration file.

Add a special message when user refuses to save a temporary file, but still tries to translate it.
Fix an ordering bug where the editor will try to remove the file from the QFileSystemWatcher after deleting the QString which holds the file path, causing access after memory was deleted.